# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PIM flow API Documentation",
        "description": "APIs documentation of PIM flow",
        "contact": {
            "email": "sysadmin@horecastore.org"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
        "/api/attributes": {
            "post": {
                "tags": [
                    "Attributes"
                ],
                "summary": "Create a new attribute",
                "description": "Creates a new attribute with name, code, and type.",
                "operationId": "48e2284fe9ed7ecdc495aa7191e74084",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "code",
                                    "type"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Color"
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "color"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "text"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/attributes/{attribute_id}": {
            "get": {
                "tags": [
                    "Attributes"
                ],
                "summary": "Get attribute details",
                "description": "Fetches attribute details based on the given attribute ID.",
                "operationId": "256ea2cae55f5b5f0b4c42f3c118a0b2",
                "parameters": [
                    {
                        "name": "attribute_id",
                        "in": "path",
                        "description": "ID of the attribute",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/attributes/{id}": {
            "put": {
                "tags": [
                    "Attributes"
                ],
                "summary": "Update an existing attribute",
                "description": "Updates an attribute's details, including name, code, type, required status, and validations.",
                "operationId": "cccd28abb387a54a131aab9ca4cc5e90",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the attribute to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "code",
                                    "type",
                                    "is_required",
                                    "validations"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Size"
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "size"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "dropdown"
                                    },
                                    "is_required": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "validations": {
                                        "properties": {
                                            "min": {
                                                "type": "integer",
                                                "example": 250
                                            },
                                            "max": {
                                                "type": "integer",
                                                "example": 260
                                            },
                                            "required": {
                                                "type": "boolean",
                                                "example": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Attribute updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Attribute updated successfully"
                                        },
                                        "attribute": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Size"
                                                },
                                                "code": {
                                                    "type": "string",
                                                    "example": "size"
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "example": "dropdown"
                                                },
                                                "is_required": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "validations": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "example": [
                                                        "required",
                                                        "max:255"
                                                    ]
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": {
        "BearerAuth": {
            "type": "apiKey",
            "in": "header",
            "name": "Authorization"
        }
    }
}